controller helper methods added
