+++ spec/controllers/likes_controller_spec.rb
@@ -1,15 +1,11 @@
 require 'rails_helper'
+require_relative '../support/controller_test_helpers_spec.rb'
 
 RSpec.describe LikesController, type: :controller do
   let(:test_post) { FactoryBot.create(:post) }
   let(:user) { FactoryBot.create(:user) }
 
   describe "POST #create " do
-    def create_like
-      allow(controller).to receive(:current_user).and_return(user)
-      post :create, params: { post_id: test_post.id }
-    end
-
     it "responds with 302 if post created" do
       create_like
       expect(response).to have_http_status(302)
@@ -22,11 +18,6 @@ RSpec.describe LikesController, type: :controller do
 
   describe "DELETE #destroy " do
     let!(:like) { FactoryBot.create(:like) }
-
-    def delete_like
-      allow(controller).to receive(:current_user).and_return(user)
-      delete :destroy, params: { post_id: like.post.id, id: like.id }
-    end
 
     it "responds with 302 if post deleted" do
       delete_like
